+++ R/imports.R
@@ -1,5 +1,7 @@
 #' @importFrom dplyr "%>%"
 #' @importFrom dplyr as_tibble bind_rows mutate
+#' @importFrom rlang .data
+#' @importFrom stats setNames
 #' @importFrom utils download.file glob2rx head type.convert
 #' @export
 dplyr::`%>%`