+++ R/rmarkdown_options.R
@@ -8,7 +8,7 @@
 #' @param break_on_error should an error in the R code execution interrupt the rendering or should rendering continue, defaults to FALSE
 #' @param ... all other arguments passed to [rmarkdown::html_document()]
 #'
-#' Custom rmarkdown input format options based on the standard [rmarkdown::html_document()], but with options that you can specify. Find the format options here in the pandoc documentation: http://johnmacfarlane.net/pandoc/demo/example9/pandocs-markdown.html
+#' Custom rmarkdown input format options based on the standard [rmarkdown::html_document()], but with options that you can specify. Find the format options here in the pandoc documentation: <https://pandoc.org/MANUAL.html#pandocs-markdown>.
 #' Pandoc's enhanced version of markdown includes syntax for footnotes, tables, flexible ordered lists, definition lists, fenced code blocks, superscript, subscript, strikeout, title blocks, automatic tables of contents, embedded LaTeX math, citations, and markdown inside HTML block elements or spoken in options: +escaped_line_breaks, +header_attributes, +yaml_metadata_block, +auto_identifiers, +implicit_header_references, +blank_before_blockquote, +fenced_code_blocks, +fenced_code_attributes, +line_blocks, +definition_lists, +startnum, +fancy_lists, +pipe_tables, +pandoc_title_block, +intraword_underscores, +strikeout, +superscript, +subscript, +tex_math_dollars, +raw_html, +markdown_in_html_blocks, +implicit_figures, +footnotes, +inline_notes, +citations.
 #' The current default rmarkdown additions to Pandoc's enhanced markdown are: +autolink_bare_uris, +ascii_identifiers, +tex_math_single_backslash, -implicit_figures. 
 #'
@@ -168,7 +168,7 @@ formr_render_commonmark = function(text) {
 formr_render = function(text, self_contained = FALSE, ...) {
   fileName = rmarkdown::render(input = write_to_file(text, 
     name = "knit", ext = ".Rmd"), output_format = formr::markdown_hard_line_breaks(self_contained = self_contained, 
-    fragment.only = FALSE), clean = T, quiet = T, ...,)
+    fragment.only = FALSE), clean = TRUE, quiet = TRUE, ...)
   fileName
 }